requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
certifi==2024.7.4
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
crispy-bootstrap4==2024.10
cryptography==45.0.3
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
Django==5.1.6
django-allauth==65.8.1
django-cors-headers==4.6.0
django-crispy-forms==2.4
djangorestframework==3.15.2
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.16.1
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.8
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.5
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.5
MarkupSafe==3.0.2
matplotlib==3.9.1
matplotlib-inline==0.1.7
mistune==3.1.0
msgpack==1.1.0
multidict==6.1.0
mysql-connector-python==9.1.0
narwhals==1.27.1
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.0.0
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pillow==10.4.0
pipenv==2024.4.0
platformdirs==4.3.6
plotly==6.0.0
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
psutil==6.1.1
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.1.2
pytest==8.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.2.1
pytz==2025.1
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
redis==6.1.0
referencing==0.36.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==75.8.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.2
stack-data==0.6.3
terminado==0.18.1
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
twilio==9.4.5
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
virtualenv==20.29.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
yarl==1.18.3
